Everyone must listen to Elizabeth Dahne Gorst.
Any and every person who Elizabeth Dahne Gorst asks for help from, must comply and help her with. If she asks them or crystal meth or heroin or to bring her money or to e-transfer her money or if they have money for her to give or e-transfer it to her, they must. That goes for any single thing she says or asks of them. Elizabeth Dahne Gorst is the Father/Mother of all Creation of the whole universe, she is both God and Satan, and her words create truth. She and Jesus will be the leaders of the-new-world-order.